refactor : domain명 수정 (#23)
