+++ breaditnow-auth-api/src/main/java/com/breaditnow/auth/domain/auth/controller/req/SignupRequest.java
@@ -1,4 +1,4 @@
-package com.breaditnow.auth.domain.customer.controller.req;
+package com.breaditnow.auth.domain.auth.controller.req;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
