+++ breaditnow-auth-api/src/main/java/com/breaditnow/auth/domain/auth/service/strategy/SignupStrategy.java
@@ -1,6 +1,6 @@
-package com.breaditnow.auth.domain.customer.service.strategy;
+package com.breaditnow.auth.domain.auth.service.strategy;
 
-import com.breaditnow.auth.domain.customer.controller.req.SignupRequest;
+import com.breaditnow.auth.domain.auth.controller.req.SignupRequest;
 import com.breaditnow.auth.global.security.Role;
 
 public interface SignupStrategy {
